swm: wft/kernel_snap -- handle snap only respins
When we have to respin a snap we have to tag it again and with a new
tag.  The contained kernel will not change version and the version it
will express in the store will not necessarily change.  Support an
additional `.N` suffix on the main version for the tag.

Signed-off-by: Andy Whitcroft <apw@canonical.com>