cranky: update-snap: rename kernel.yaml to kernel-info.yaml
kernel.yaml is already a thing that kernel snaps can specify, it gets
copied into meta/ directory, and parsed by snapd when installing
kernel snaps. It has an expected structure & keys as defined in
snapd. Whilst we can use kernel.yaml for our internal purposes, the
keys we use may clash with snapd's interpretation of them. It is best
if we use a unique/different name for kernel-info.yaml solely for our
internal/swm purposes.

Any existing updates of kernel snaps will need to be checked if they
already have a snapd or swm kernel.yaml, and fixed up appropriately.

Fixes: d439b4309b ("cranky: update-snap: Migrate to use kernel.yaml")
Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>