swm: snap -- fix various channel-map caching issues
We have a number of issues with the caching of the snap-store data for
the snaps.The caching in `channel_revision` is just broken, but
luckily we always call `channel_version` first which does it correctly
and causes the former to function correctly.

Separatly we are not caching the channel-map for the last-published use
case.  This means that the scan we do to determine our snap tracker
disposition could be out of sync with the scan we record for monitoring
purposes.  This creates a race window in which the disposition could
change without us noticing.

Fix all of this by directly caching the channel-map at the `channel_map`
level.

Signed-off-by: Andy Whitcroft <apw@canonical.com>