info/kernel-series: route all stable/devel kernels with signing via a signing PPA
We are now routing everything via the signing PPAs for consistency even
if it does not have drivers (lrm et al).  We are currently doing this
via a hack in canonical-signing bot.  Switch routing formally in
kernel-series to make this the norm.  This cleans up swm tracking during
the signing dance.

Signed-off-by: Andy Whitcroft <apw@canonical.com>