swm: context,manager -- add a new context object and seed
Add a new context object which can contain arbitary global values.  This
is intended to hold pointers to common contextual data like KernelSeries
data.  Things which are horribly expensive to generate, and which we do
not want to have to continually pass from object to object in case they
are needed.

Also add an initialisation function in manager to seed the values we are
going to use; none are seeded initially.

Signed-off-by: Andy Whitcroft <apw@canonical.com>