info/kernel-series: pull out shared assets
We make heavy use of the &anchor *anchor alias nodes to share data like
the routing table.  The yaml standard requires the anchor definition be
before being referenced via an alias.  This leads us to a situation
where we have to pick up these assets and move them when we add a new
series.

Long ago we added a defaults section for all series, but never made use
of it.  This pseudo series 'defaults' is layered underneath the main
series definitions.  This did not work well as these are often not
defaults at the right level.  Also to add new default behaviour we have
to modify and deploy the library.  Finally, we do not want necessarily
to default a value, we commonly want most of one key, and half of
another; which we have implemented using anchor/alias notation.

Pull out the assets into this new 'defaults' section.  We intentionally
group these further under an 'assets' key.  This allows us to add futher
sections later.  This also provides compatibility for any out-of-date
versions of the library which still applies the defaults section as an
underlay as the 'assets' key will be ignored.

Signed-off-by: Andy Whitcroft <apw@canonical.com>