swm: source-blocks -- switch to a common source_blocks_present and add jira
We are adding support for jira generated blocks.  These are going to
come in via tagging.  Pull out the three copies of the tag analysis and
commonise it on the WorkflowBug objects.  Merge up the old manual
kernel-block-source and kernel-block into this and share to all
locations.  Include a clear reason so we can ensure the reasons shown in
the dashboard remain nice and specific.

Signed-off-by: Andy Whitcroft <apw@canonical.com>