info/kernel-series.yaml: Add focal/linux-azure-cvm
This is a temporary kernel that will be used as base for the uc20-efi
like kernel linux-azure-fde while all the necessary changes are not
applied to our regular azure kernels.

Since it will only be used as a base for azure-fde, there's no need
for a meta or a signed package. ADT and RT are also marked as invalid
and those tests should be executed with linux-azure-fde instead.

There's also no need for promoting this kernel past the build PPA, so
mark the corresponding task as invalid too.

Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>