bug2issue: Add a script for creating Jira issues from LP bugs
This is similar to bug2card but for creating Jira issues. At the minimum
a Jira component, Jira username (email address) and LP bug number need
to be provided. Optionally, a Jira API token, project name and/or desired
issue status can be specified.

Authentication with Jira is performed using the provided Jira username and
an API token which needs to be generated manually in Jira beforehand. Also,
the provided component name obviously needs to exist in Jira.

If no API token is provided on the commandline, the script pulls it from
the password store using the pass name 'warthogs.atlassian.net/<username>'
(see https://www.passwordstore.org).

If the --dry-run option is specified, all the necessary data is pulled from
the LP bug, authentication with Jira is performed, some relevant data is
printed out but no issue is created.

An example for creating an issue on the 'Raspberry Pi board':
$ bug2issue -c 'Raspberry Pi' -u juerg.haefliger@canonical.com 1949278

And the resulting issue:
https://warthogs.atlassian.net/browse/KERN-598

Signed-off-by: Juerg Haefliger <juergh@canonical.com>