info/kernel-series: jammy/realtime: mark as development, v5.15 and derived from jammy
The jammy/realtime kernel should be rebased on latest generic kernel,
not from impish/realtime.  There is nothing shared with impish/realtime.

The kernel will be based on v5.15 so currently on unstable. Mark it as
development and still not supported (no SRU even), however the comment
does not apply anymore as there will be active work on it from realtime
engineers.

Signed-off-by: Krzysztof Kozlowski <krzysztof.kozlowski@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>