swm-jira-sync: only assign issues which are not yet assigned
We want to assign things to work on in Jira without changing the
nominal ownership.  Such as cranking a kernel for someone else.
Here we will assign it to us and then unassign it.  By switching
(back) to only assigning when there is no assignee we will allow
this use case.

Signed-off-by: Andy Whitcroft <apw@canonical.com>