swm-jira-sync: use variant tags to separate newly opened and closed trackers
We care to separate newly opened and closed trackers.  Both of these
appear as jira issues without a corresponding tracker.  We used to leave
the issue in In Backlog until we saw a tracker, but this leads to
potential for the tracker to leave In Backlog on the wrong workflow; so
now we wish to shove those into Kernel Not Reader sooner, and need a new
way to separate them.  If we have previously seen a tracker we would
have labelled this issue, check for those tags to separate these cases.

Signed-off-by: Andy Whitcroft <apw@canonical.com>