cranky: review-master-changes: Fix script for derivatives of derivatives
Update `cranky review-master-changes` to properly work with
derivatives of derivatives (ie backport kernels).

For derivatives of derivatives, since the base kernel is rebased for
each new version, that means that in git the current version of the
base kernel doesn't descent from its previous version.

Because of that, instead of listing the base commits from what was
fetched, find the correspoding commit in the current branch instead
and use that as the starting point.

Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>