swm: source-blocks -- ensure mirror flags are updated in all cases
We care very much that the mirror flags are updated to match the current
state when we save the tracker.  External processing is relying on those
being accurate to optimise handling.  Ensure we write them back on every
status change regardless of whether we scanned against those tags.  Note
that if the flag is unchanged then swm-properties remain unchanged and
their save will be avoided.

Signed-off-by: Andy Whitcroft <apw@canonical.com>