swm: package -- detect packages which used to belong to us
When we "pull-back" a tracker and reuse it (an in-tracker respin) such
as is done when the kernel partially fails to build and was never
published, we generate orphaned binary package versions from these
respun packages which (seemingly) do not belong to anyone.  These
prevent us from promoting.  Currently we only have the big hammer tags
to force this.

Add a new swm-properties field "versions-replace:" which provides for a
per package-type list of versions which this tracker is responsible for
in addition to its current package versions.

For example if we respin lrm on a kernel from 5.15.0-1.1 to 5.15.0-1.1+1
we would record the old version:

    versions-replace:
        lrm: [5.15.0-1.1]

Signed-off-by: Andy Whitcroft <apw@canonical.com>