cranky: review-master-changes: properly find base version with ~suffix
If base version of a derivative is a backport with ~suffix, the first
cranks will have a changelog entry containing '~', however the tags have
it replaced with '_'. For example bionic:ibm-gt-5.4 is based on
bionic:hwe-5.4, therefore first changelog entries have
5.4.0-91.102~18.04.1 and tag Ubuntu-hwe-5.4-5.4.0-91.102_18.04.1.

Replace the '~' with '_' when looking for tags to cover such case.

Signed-off-by: Krzysztof Kozlowski <krzysztof.kozlowski@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>