cranky: dput-sources: also look up files on current directory
Some workflow move files to be uploaded to a given directory, then review
and upload them from there. This change keeps looking first at the parent
directory to the source tree, but if the file to be uploaded is not found
there, it is looked into the directory where the command was started.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>