kernel-stable-dashboard: handle incoming top-level trackers key
We are adding a top-level trackers key to the SwmStatus data.  This
allows us to add things like lists of active cycles and the like to the
data without perturbing users.  The SwmStatus object has had handling
for this change for some time.  Add similar handling to the open-coded
access in the dashboard.

Signed-off-by: Andy Whitcroft <apw@canonical.com>