ktl/kernel_series: clean up 'defaults' handling and elide assets
We are currently cheating in kernel-series using an assets key within
the defaults section to hold YAML fragments we are going to reuse.  This
leads us to carry those defaults in all series data blocks even though
they are never used. Elide them formally from the data.  Also take the
oppotunity to clean up the handling of defaults which we pass around
pointlessly.  Use them directly from the KernelSeries object.  Finally
prepare for the possibility that we may wish to have defaults at other
levels by using the 'series' key within defaults if it is present; thus
retaining backwards compatibility.

Signed-off-by: Andy Whitcroft <apw@canonical.com>