cranky: chroot: add builder-extra ppa to apt source list
The builder-extra ppa is added as a dependency to all of our official
build ppa's and can hold for example packages update which has not made
it to -security yet. Some of these packages can be required during the
crank, so add this ppa to the cranky chroot sessions.

For that we also need to add the ppa's public signing key, which is
kept locally on builder-extra.asc. This is the ASCII armored of
ppa:canonical-kernel-team/builder-extra signing key (fingerprint:
110E21D8B0E2A1F0243AF6820856F197B892ACEA).

Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>