swm: package -- consider duplicate package ownership by package
Currently we consider duplicate trackers as owning binaries in a pocket
if it owns all of the outstanding binaries.  With partial respin
trackers we may have one duplicate tracker owning the main, signed, and
meta package, while a second duplicate owns lrm, lrg, and lrs.

Consider each package separately.  As long as someone we believe we
replace owned a package treat it as ours.

Signed-off-by: Andy Whitcroft <apw@canonical.com>