swm: wft/promote-to-proposed -- ensure we have maximal delta on multi-hop transfers
When we are using a signing PPA the PromoteToProposed object is
mediating a multi-hop movement.  In this scenario we care about all of
the packages we will need to copy out of the initial PPA into later
ones, even of some of those are embargoed and will subsequently be
eliminated.  Accumulate the union of each delta set.

Signed-off-by: Andy Whitcroft <apw@canonical.com>