cranky/update-snap: Handle version but no source in snapcraft.yaml
While working on other issues when producing UC20 snaps we ended up in a
(likely temporary) state where we still have a version key in
snapcraft.yaml but no source (URL) key. This was not expected to happen
and so update-snap craters if it cannot update both. Add code to make
that optional.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>