cranky: chroot: do not add builder-extra to esm kernels
The current method used to add the builder-extra ppa to the apt source
list doesn't work on xenial and breaks the config. The PPA used to build
the ESM kernels doesn't have any extra dependency, so just change the
code to not add this PPA to trusty and xenial for now instead of fixing
it for xenial and older.

Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>