swm: automated-testing -- report SEED-AWOL and similar as manual-intervention required
When the adt-matrix has no seed data for a kernel, it cannot sensibly
determine regression status.  In this case it reports the testing as
ongoing, which we calmly report as testing in-progress.  This is easy to
miss and assume testing will complete.  Detect this state and convert it
into TaskError so that it is front and center.

Signed-off-by: Andy Whitcroft <apw@canonical.com>