swm: wft/regression_testing -- align test message generation to its task handler
Currently we emit a regression-testing request when promote-to-proposed
is complete.  However, this is an integral part of the implementation of
regression-testing.  Pull this message generation into the
RegressionTesting object.  This brings message generation into line with
boot-testing; and corrects this layering violation.  Finally it allows
us to override this testing by marking it invalid.

Signed-off-by: Andy Whitcroft <apw@canonical.com>