swm: promote-to-release -- separate britney owned packages from the rest
When we have a live promote-to-release we will need to handle it
differently if it is going to the main archive or elsewhere.  For the
main archive britney is in charge.  Separate promotions to the main
archive by moving the task Triaged and report those in that state as
with britney.

Signed-off-by: Andy Whitcroft <apw@canonical.com>