swm: promote-to-proposed -- use Triaged to indicate pre-approved new-review
When we pre-approve a new-review we can indicate this to swm by moving
the promote-to-proposed to Triaged.  This will avoid it showing up on
the manual review list.

NOTE: this does not cause the new-review to be actioned, this requires
the reviewer to submit a pre-approved signing request to
canonical-signing-bot.

Signed-off-by: Andy Whitcroft <apw@canonical.com>