swm: bug -- add a new private properties store
We have a mix of public and private properties in swm-properties.  It is
hard to quickly determine which you might wish to change and which
really should be left alone.  Create a new private_props to store any
such properties in.  Create a new top level key swm-properties key `~~`
to hold anything which is intended to be internal only; this key is
chosen as it represents cleanly in yaml (no quoting required), and sorts
to the bottom.  Later patches will move relevant data over to this key.

Signed-off-by: Andy Whitcroft <apw@canonical.com>