swm: package -- add support for a prepare-id
Add a new prepare_id accessor which encodes the entire valid version
set.  This will change if any of the packages in the build-set changes.
We can compare the current prepare-id to one recorded when (say) we
commited an sru-review, and if it changes know that the review must be
recinded.

Signed-off-by: Andy Whitcroft <apw@canonical.com>