cranky: fix: do not overwrite existing Ubuntu.md changes
Currently cranky fix always re-create the Ubuntu.md file overwriting any
manual changes made to it. Skipping making changes to the file when it
already exists is not enough because new derivatives would not have the
file automatically updated for the first time. So try to be a bit
smarter, by not touching the file on master kernels when it already
exists, and for derivatives/backports making any change only if there
isn't any commit applied locally.

v2:
  - Handle derivative kernels which are not rebased (e.g. uc*-efi)

Acked-by: Cengiz Can <cengiz.can@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>