swm: package -- record unaccounted for packages by type/version
After scanning a pocket to ensure we can replace everything in it and
we find it is not clear accumulate and report the unaccounted for
packages.  These are exposed as a new synthetic reason;
packages-unaccounted.  They are exposed in a form such that they could
be pasted directly into versions-replace once reviewed and accepted as
ours.

Signed-off-by: Andy Whitcroft <apw@canonical.com>