swm: new-review -- handle preapproval via Triaged
When we wish to preapprove the new-review (at the same time as
sru-review) we have been moving new-review directly to Fix Released.
However when the new-review is then scanned it immediatly is recinded
back to New because we have no cached prepare-id.  Later it will move to
Confirmed in the normal way; leading to confusion.

Instead have the pre-approver move the new-review to Triaged to indicate
this form of approval.  When we find it in this state move directly to
Fix Released, and promote the sru-review clamp to new-review; this is
the version they reviewed so it make sense to share this clamp.

Signed-off-by: Andy Whitcroft <apw@canonical.com>