cranky: review: print previous version
When running cranky review, it will lookup for a previos version to compare to.
Its output is not verbose, so the user may not even be aware of which version
it is comparing against until it is finished and the resulting debdiff is
looked at.

Print the version it is comparing against, so the user is able to stop before
its completion, in case it is the wrong version.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Cengiz Can <cengiz.can@canonical.com>
Acked-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>