kernel-series: make lowlatency-hwe-5.15 non -edge
As part of the roll of the default hwe kernel in focal from hwe-5.13 to
hwe-5.15, linux-lowlatency-hwe-5.15 also needs to have its variants
updated. For hwe-5.13 the lowlatency flavors were built by the main
package, so the counterpart of this change is the previous commit
where hwe-5.13 variant was set as 'eol-hwe-20.04'.

Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>