swm: wft/* -- follow changes to is_derivative_package
is_derivative_package is no longer derfinative as we may choose to
ignore the propered master-bug if it is gone.  This will not be
discovered until we instantiate the parent tracker.  Use the validility
of the master_bug to determine this only when we are about to check the
tracker internals and instantiate it anyway.

Signed-off-by: Andy Whitcroft <apw@canonical.com>