info/kernel-series.yaml: Re-enable test for focal/linux-azure-cvm
Regression and boot tests do not work anymore with
focal/linux-azure-fde because we had to remove the postinst from its
meta, that used to replace grub with the kernel EFI binary.

Until we are able to spin Azure CVM instances to test linux-azure-fde
directly, keep these tests enabled for azure-cvm. I'm keeping the
regression test task for linux-azure-fde for now, because I will use
it to temporarily block its promotion until I have boot tested it
manually on a CVM instance.

Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>