swm: promote-to-proposed -- report missing builds more agressively
Move to waiting on :prepare-packages too so we do not start checking on
builds before they are complete (such as could happen when boot-testing
is invalid or overriden).  Now we are not relying on the completion
test until they _should_ be complete, report agressivly when they are
not.  This catches the case where a build is erroneously marked complete
(at the prepare-package level).

Signed-off-by: Andy Whitcroft <apw@canonical.com>