swm-jira-sync: handle duplicate Jira issue tags
If someone makes a second Jira issue for a kernel we will blindly detect
the tracker tag as missing and add a new one.  This leads us to have
more than one tracker issue tag (with different issue numbers) and
non-deterministic behaviour.  Detect a non-matching issue tag and remove
it.  This will over time consume all incorrect tags; one per run.  In
principle we would never have more than one when controlled by
swm-jira-sync so this slow correction is not an issue.

Signed-off-by: Andy Whitcroft <apw@canonical.com>