info/kernel-series: sort out routing for linux-realtime
We wish to stop updating the old `proposed` PPA as people using that
should now be using UA to access the official mirrors.  To this end make
a new as-proposed PPA and move the as-proposed there.

We only wish to publish linux-realtime in the LTS releases (jammy only).
Add new realtime-interim routing for interim releases which route the
final binaries to a private PPA for tracking and testing.

Signed-off-by: Andy Whitcroft <apw@canonical.com>