swm-worker: initial AMQP based worker
Add an initial framework for an AQMP based worker.  This initially
supports the `shank` command only.  The plan is to pull all of the
various operations (instantiate, dependent scan etc) into this core
worker.  With this in place we can then run multiple workers to improve
responsivness and improve failure tollerance.

Signed-off-by: Andy Whitcroft <apw@canonical.com>