swm-worker: pass in the expected scan time
Some operations can lead to duplicate shank requests being queued.  For
example if we enqueue all dependant trackers when a parent changes we
cannot tell if they have previously been queued or note.  This leads to
a high probability of duplicate requests in the queue.  Add the current
scan time to the request, if when we go to scan the tracker it has
already been scanned this scan time will missmatch and we can simply
drop the duplicate request.

Signed-off-by: Andy Whitcroft <apw@canonical.com>