swm-jira-sync: handle adding a jira issue number where none is present
When a tracker without issue is detected swm-jira-sync will actually
make a new issue for it and attempt to add the required tracker tag.  We
will correctly indicate the desire to add the new issue link, and
indicate there is None to delete.  Currently this will lead
swm-jira-sync to crater.  Detect this and avoid trying to remove the old
link.

Signed-off-by: Andy Whitcroft <apw@canonical.com>