ktl/kernel_series: add package type relationships
Add a new set of relationships between packages in a source (see below
for a full set).  These allow tooling to identify ordering, privacy, and
signing requirements for the various packages.  We also add a new
package-relationships table with names relationship data.  This allows
sources to indicate which semantics they have; for example for a source
converted to have a "linux-generate" package.

Each package has a set of associated keys as below:

adjunct: <true/false> -- the package is an adjunct package, a package
which must be uploaded to the private build PPA;

ancillary-for: <ptype> -- the package type of the package we are an
ancillary package for, the package in which this package is contained
and by which it is versioned;

depends: <ptype> -- the package type of the package we depend on, one
which much complete building before we can build;

signing-to: <ptype> -- indicates a package generates a custom signing
tarball and points to the package type which will ultimatly consume the
signed objects; and

signing-from: <ptype> -- indicates a package which consumes a custom
signing tarball and points to the package which generates it.

Signed-off-by: Andy Whitcroft <apw@canonical.com>