ancillary-bot: lookup use librarian to grab source earlier
We currently wait for sources to move Published so we can grab them via
the pool.  Switch to using our launchpad connection to convert the Urls
to tokenised librarian links and use those.  This allows us to prepare
ancillaries for packages in Pending and therefore get the ancillary
packages uploaded a publisher-cycle earlier.

Signed-off-by: Andy Whitcroft <apw@canonical.com>