info/kernel-series: update jammy bluefield kernel entry
This kernel has been renamed to linux-bluefield from linux-nv-bluefield so
update the kernel entry name, names of packages and URLs for repos to match
the change. Also this kernel will be a part of the usual SRU cycle (starting
from the 2022.09.19 one) so change 'supported' to 'true', 'development' to
'false' and remove no longer needed special 'routing' information.

Signed-off-by: Bartlomiej Zolnierkiewicz <bartlomiej.zolnierkiewicz@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>