swm: package -- record the launchpad source object
Record the launchapad SPPH record for the source we are examining on a
per-pocket basis, even for pockets which do not contain our package.
This will allow us to use these for publishing "change" detection; if we
have the wrong version published, if that publishing record changes
there is reason to recheck that publication.

Signed-off-by: Andy Whitcroft <apw@canonical.com>