swm: new-review -- differentiate approvals prior to builds complete
Differentiate between approves before and after builds complete.  In the
former we cannot have reviewed actual binary package names for New
packages.  Expect the reviewer to review to Fix Committed in this
degenerate case.  We will then check for New packages and if none are
found automatically transition to Fix Released; otherwise moving back to
Triaged to get a human review.

Signed-off-by: Andy Whitcroft <apw@canonical.com>