cranky: add list-repos sub-command
This adds a cranky command which can be used to query the repo urls and
main branch associated with each package for one or more handles.
Optionally, packages may be restricted by type e.g. signed, meta, etc.
This is intended to help find repo urls and main branch names without
having to dig through the kernel series yaml. Output is standardized to

<codename> <package type> <web url | git clone url> <branch name>

Signed-off-by: Cory Todd <cory.todd@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>