cranky: fix kinetic bash completion
Starting with bash-5.2, as found on Kinetic, cranky bash completion is
broken when attempting to complete handles. This is due to a change in
how the ':' character is handled and escaping it is no longer adequate
for our purposes.

The fix is to rely on two functions from bash-completion which we will
assume are present if you are expecting bash completion. So instead of
clobbering COMP_WORDBREAKS which is global, we complete a word up to a
':' then separately complete after the ':'. This is similar to how NPM
and Maven have fixed their bash completion tools.

If bash-completion is not installed then cranky complete will silently
fail.

Signed-off-by: Cory Todd <cory.todd@canonical.com>