cranky: crl/cycle_ref: add helper for working with git refs
This class is used for listing and generating git refs as used by our
cycle tag mapping scheme. The algorithm is pretty straightforard. For
each kernel crank we inject the cycle into $debian/tracking-bug. By
finding these commits we can associate a tag and a cycle from which a
tree of refs/swm/$handle/$branch-$cycle can be generated.

Change-Id: I775150d48ca03c5462b5f6e0ee469d571b552bea
Signed-off-by: Cory Todd <cory.todd@canonical.com>
Reviewed-on: http://ckctreview.internal/r/c/kteam-tools/+/1262
Acked-by: Jose Ogando Justo <jose.ogando@canonical.com>
Signed-off-by: Cory Todd <cory.todd@canonical.com>