swm: package -- copies result in publications with no builds
When we copy a source with binaries between series we end up with binary
publications with no builds.  This trips the build missmatch code
because we have more publications than builds require.  Our actual
desire is to confirm that every build has publications.  Switch to
checking if the build architectures are not a subset of the publication
architectures.

Signed-off-by: Andy Whitcroft <apw@canonical.com>