mainline-builds: cod-ubuntu-config-trigger -- process 100 new tags at a time
If we're falling behind and there's more than 100 new tags pulled
from the repos, the script just bails and never queues any new
config builds:

  cod-ubuntu-config-trigger: 3648 new tags is greater than 100 not processing.

Fix that by processing the first 100 of the new tags. Depending
on the number of new tags, it will take a couple of runs to
catch up again.

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Acked-by: Cory Todd <cory.todd@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>