info/kernel-series: rename jammy/linux-intel-iotg-realtime to linux-intel-iot-realtime
Intel IOTG has dropped the 'G' from their name.  So we want the kernel
to go by just linux-intel-iot-realtime.

The non-realtime kernel will also be changed, but that is more involved
as it is an existing kernel.  The realtime varient of the intel IOT
kernel has not been upload yet, and I want to get this change done
before I upload the intial version.

Signed-off-by: Philip Cox <philip.cox@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>