tests: move tests out to their own directory
We really want to use relative imports in our API accessors.  To do this
we will need to add a ktl link where our tests are and we would rather
not have that in the top level of the library.  Therefore move all of
the tests into their own directory and instantiate it there.

Signed-off-by: Andy Whitcroft <apw@canonical.com>