cranky: review: improve meta versioning
When normalizing the versions, it is assumed that dash is the separator
between the upstream version and the ABI. For meta packages, it is a dot.

However, there are still instances on the meta packages DSC and control
files, where the version with the dash as a separator is referenced,
specially on package relationships. So, also replace those instances.

Change-Id: I1d44e25e32fe388477135fa00a9eec1662408abf
Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Reviewed-on: http://ckctreview.internal/r/c/kteam-tools/+/1964
Acked-by: Cory Todd <cory.todd@canonical.com>
Acked-by: Andrei Gherzan <andrei.gherzan@canonical.com>
Signed-off-by: Cory Todd <cory.todd@canonical.com>