mainline-builds: cod-mainline-core -- Reconstruct etc/kernelconfig
For mainline builds we pull the packaging scripts and the kernel
configs from different points in time. With the recent changes to
annotations-only, we can end up in a situation where we have
old-style configs but new-style packaging scripts. Luckily, the
old script are still around and are used in that case but they need
etc/kernelconfig which might no longer exist. Reconstruct it, if
that's the case.

Change-Id: I1704ba16744dd2586a8c5e98d3f8d1f11180c0ff
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Reviewed-on: http://ckctreview.internal/r/c/kteam-tools/+/1671
Acked-by: Cory Todd <cory.todd@canonical.com>
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>