info/kernel-series: initial updates for key based routing
Switch to a new routing model.  Create new routing tables which indicate
routing stream and which key is in use.  This allows us to say the
following to select signed routing:

    source foo:
       routing: ubuntu/2

If we do not specify signed routing unsigned routing is used by default.
In this way we need to take explicit action to select a new key.

Reconstruct the various tables using a layering model which separates
security/build/release routing sections and layers them.

Signed-off-by: Andy Whitcroft <apw@canonical.com>