info/kernel-series.yaml: Invalidate cert-testing on focal:linux-iot
Certification Testing has not been engaged to test .deb
focal:linux-iot and they appear to have to recollection that any
cert-testing is supposed to be done.

All prior trackers got released with invalidated certification
testing.

If certification testing is supposed to be done on this kernel, please
engage with ~Devices Certification, provide devices to test, ensure
that cert-testing is setup, and then revert this patch.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>