info/kernel-series: add build-private routes for ibm-gt and realtime
Add build-private routing for ibm-gt and realtime.  Neither is used
as these do not have LRM currently but it is good practice to be
prepared for this to change.  It also greatly simplifies the backend
configuration to do so.  As each uses private build routes they are
valid as build-private entries.

Signed-off-by: Andy Whitcroft <apw@canonical.com>