info/kernel-series: only tested arches get stable snaps
armhf pc-kernel is unused (snapstore kpis) and in fact impossible to
use, as it neither supports EFI (not that it exists on armhf) nor are
there any boards usable with it via u-boot. On armhf, one typically
uses the pi-kernel or any of the other custom kernels. Close the armhf
channel of the pc-kernel.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>