swm: new-review -- check Ubuntu-Signing-Compatible is consistent
When we pre-approve a new-review it is on the expectation that there
will be no actual new packages; any deviation from this triggers swm
to refer the new-review back.

Add an additional check to check that the Ubuntu-Signing-Compatible
data is consistent from the source to the produced binaries.  If they
are not consistent also refer the new-review back.

Signed-off-by: Andy Whitcroft <apw@canonical.com>