adt-matrix: Treat superficial-PASS tests as PASS
BugLink: https://bugs.launchpad.net/bugs/2007969

Since autodep8 0.26, the dkms packages are marked as superficial. This
triggers autopgktest to treat these similarly to SKIP. This means that
a test suite that "passed" but only with superficial tests, will be
surfaced with exit code 8:

8 no tests in this package, or all non-superficial tests were skipped

From the perspective of the ADT matrix, we want to maintain the old
behaviour where even superficial passes would be considered as passed
tests. Given that we now can't distinguish this purely from the exit
code of autopkgtest, we parse its summary file.

Change-Id: Ia52c5bcda866d295e958c731012a7a27df173691
Signed-off-by: Andrei Gherzan <andrei.gherzan@canonical.com>
Reviewed-on: http://ckctreview.internal/r/c/kteam-tools/+/1818
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>