adt-matrix: adt-lp-binaries -- improve rebuild algorithm
Improve performance of mapping new kernel/version/flavours to their
corresponding sources.  Do this by caching the latest complete source
publication and querying using the "created_since_date" mode for
getPublishedSources.

Take this oppotunity to better express which keys are internal cache
keys (i-*) and which are for direct use.

Signed-off-by: Andy Whitcroft <apw@canonical.com>