cycle_ref: read debian.env from tag
When creating cycles references, the tracking-bug should be read from the
${DEBIAN} directory, the one defined in debian/debian.env.

While checking for that file in the git repository, read_cycle would not check
for the debian/debian.env file on the tagged commit, but on the checked out
tree, which may even belong to a different source package.

Also, when parsing that file, debian_env variable contained a string, while a
list was expected. As a result, debian.master was always being used.

Change-Id: I3909208c52232b75b26bd74a55e44310ec343fb3
Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Reviewed-on: http://ckctreview.internal/r/c/kteam-tools/+/1952
Acked-by: Cory Todd <cory.todd@canonical.com>
Signed-off-by: Cory Todd <cory.todd@canonical.com>