cranky: checkout: allow --cycle and --crd to be used together
When checking out a CRD branch, if there is an already pushed branch in the
embargoed repo, that one will be used. Otherwise, the default remote
reference will be checked out.

When preparing a CRD on top of a given cycle, all the usual CRD effects
should be used, but the reference to be checked out in case there is not
one already should be the given cycle reference.

Change-Id: I20bf30bd76fa5a7c1472390a7512de8c3633c6ee
Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Reviewed-on: http://ckctreview.internal/r/c/kteam-tools/+/1955
Acked-by: Cory Todd <cory.todd@canonical.com>
Signed-off-by: Cory Todd <cory.todd@canonical.com>