rebase: add -o (own debian dir) option
When kernels are prepared underneath their master-next branch, one should
rebase the staged changes into the latest prepared kernel.

    Ubuntu-aws-5.4.0-1001.1
o---o
     \
      o---o---o master-next

    'Ubuntu-aws-5.4.0-1001.1
o---o
     \
      o---o---o---o Ubuntu-aws-5.4.0-1002.2

In the above scenario, 'Ubuntu-aws-5.4.0-1001.1 is the result of rebasing
Ubuntu-aws-5.4.0-1001.1 on top of the parent kernel branch. What we want to
achieve here is find out the synthetic merge base between master-next and
Ubuntu-aws-5.4.0-1002.2, which is Ubuntu-aws-5.4.0-1001.1.

In order to do that, we need to look at debian.aws/ ($DEBIAN), instead of
debian.master ($DEBIAN_MASTER).

The end result is:

    'Ubuntu-aws-5.4.0-1001.1
o---o
     \              Ubuntu-aws-5.4.0-1002.2
      o---o---o---o
                   \
                    o---o---o master-next

This should be used giving the local branch or tag with the -l option, and
should really be considered as a pumbling option, rather than cranky
porcelain. Use with care.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>